Refactor: libraries: move deprecated APIs to separate headers
This simplifies handling of deprecated APIs guarded by PCMK__NO_COMPAT.

Supported public API headers now declare only supported APIs, and include the
compatibility header when PCMK__NO_COMPAT is not defined.

The (new) compatibility headers declare only deprecated APIs, and can be
included by the internal source file defining the APIs, avoiding duplication of
declarations without exposing the deprecated APIs to all files.
