+++ include/crm/cluster/Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright 2012-2017 the Pacemaker project contributors
+# Copyright 2012-2021 the Pacemaker project contributors
 #
 # The version control history for this file may have further details.
 #
@@ -22,4 +22,4 @@ MAINTAINERCLEANFILES    = Makefile.in
 headerdir=$(pkgincludedir)/crm/cluster
 
 noinst_HEADERS		= internal.h election.h
-header_HEADERS		= 
+header_HEADERS		= compat.h
