+++ include/crm/common/Makefile.am
@@ -12,7 +12,12 @@ MAINTAINERCLEANFILES = Makefile.in
 headerdir=$(pkgincludedir)/crm/common
 
 header_HEADERS = xml.h ipc.h util.h iso8601.h mainloop.h logging.h results.h \
-		 nvpair.h acl.h agents.h ipc_controld.h ipc_pacemakerd.h
+		 nvpair.h acl.h agents.h ipc_controld.h ipc_pacemakerd.h \
+		 agents_compat.h	\
+		 logging_compat.h	\
+		 mainloop_compat.h	\
+		 util_compat.h		\
+		 xml_compat.h
 noinst_HEADERS = internal.h alerts_internal.h \
 		 iso8601_internal.h remote_internal.h xml_internal.h \
 		 ipc_internal.h output_internal.h cmdline_internal.h	\
