+++ include/crm/msg_xml_compat.h
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2004-2021 the Pacemaker project contributors
+ *
+ * The version control history for this file may have further details.
+ *
+ * This source code is licensed under the GNU Lesser General Public License
+ * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
+ */
+
+#ifndef XML_TAGS__H
+#  define XML_TAGS__H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * \file
+ * \brief Deprecated Pacemaker XML constants API
+ * \ingroup core
+ * \deprecated Do not include this header directly. The XML constants in this
+ *             header, and the header itself, will be removed in a future
+ *             release.
+ */
+
+//! \deprecated Use PCMK_STONITH_PROVIDES instead
+#define XML_RSC_ATTR_PROVIDES       "provides"
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
