+++ include/crm/pengine/pe_types.h
@@ -526,25 +526,7 @@ typedef struct pe_action_wrapper_s {
 } pe_action_wrapper_t;
 
 #ifndef PCMK__NO_COMPAT
-/* Everything here is deprecated and kept only for public API backward
- * compatibility. It will be moved to compatibility.h in a future release.
- */
-
-//!< \deprecated Use pe_action_t instead
-typedef struct pe_action_s action_t;
-//!< \deprecated Use pe_action_wrapper_t instead
-typedef struct pe_action_wrapper_s action_wrapper_t;
-//!< \deprecated Use pe_node_t instead
-typedef struct pe_node_s node_t;
-//!< \deprecated Use enum pe_quorum_policy instead
-typedef enum pe_quorum_policy no_quorum_policy_t;
-//!< \deprecated use pe_resource_t instead
-typedef struct pe_resource_s resource_t;
-//!< \deprecated Use pe_tag_t instead
-typedef struct pe_tag_s tag_t;
-//!< \deprecated Use pe_ticket_t instead
-typedef struct pe_ticket_s ticket_t;
-
+#include <crm/pengine/pe_types_compat.h>
 #endif
 
 #ifdef __cplusplus
