+++ include/crm/pengine/rules.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2020 the Pacemaker project contributors
+ * Copyright 2004-2021 the Pacemaker project contributors
  *
  * The version control history for this file may have further details.
  *
@@ -74,46 +74,7 @@ gboolean pe_eval_subexpr(xmlNode *expr, pe_rule_eval_data_t *rule_data,
                          crm_time_t *next_change);
 
 #ifndef PCMK__NO_COMPAT
-/* Everything here is deprecated and kept only for public API backward
- * compatibility. It will be moved to compatibility.h in a future release.
- */
-
-//! \deprecated Use pe_evaluate_rules() instead
-gboolean test_ruleset(xmlNode *ruleset, GHashTable *node_hash, crm_time_t *now);
-
-//! \deprecated Use pe_test_rule() instead
-gboolean test_rule(xmlNode *rule, GHashTable *node_hash, enum rsc_role_e role,
-                   crm_time_t *now);
-
-//! \deprecated Use pe_test_rule() instead
-gboolean pe_test_rule_re(xmlNode *rule, GHashTable *node_hash,
-                         enum rsc_role_e role, crm_time_t *now,
-                         pe_re_match_data_t *re_match_data);
-
-//! \deprecated Use pe_test_rule() instead
-gboolean pe_test_rule_full(xmlNode *rule, GHashTable *node_hash,
-                           enum rsc_role_e role, crm_time_t *now,
-                           pe_match_data_t *match_data);
-
-//! \deprecated Use pe_test_expression() instead
-gboolean test_expression(xmlNode *expr, GHashTable *node_hash,
-                         enum rsc_role_e role, crm_time_t *now);
-
-//! \deprecated Use pe_test_expression() instead
-gboolean pe_test_expression_re(xmlNode *expr, GHashTable *node_hash,
-                         enum rsc_role_e role, crm_time_t *now,
-                         pe_re_match_data_t *re_match_data);
-
-//! \deprecated Use pe_test_expression() instead
-gboolean pe_test_expression_full(xmlNode *expr, GHashTable *node_hash,
-                                 enum rsc_role_e role,
-                                 crm_time_t *now, pe_match_data_t *match_data);
-
-//! \deprecated Use pe_unpack_nvpairs() instead
-void unpack_instance_attributes(xmlNode *top, xmlNode *xml_obj,
-                                const char *set_name, GHashTable *node_hash,
-                                GHashTable *hash, const char *always_first,
-                                gboolean overwrite, crm_time_t *now);
+#include <crm/pengine/rules_compat.h>
 #endif
 
 #ifdef __cplusplus
